feat: add contract model and update contract service to use new model
